# contracts/openapi/local-api.v1.yaml
openapi: 3.1.0
info:
  title: Jarvis Local API
  version: 1.0.0
  description: Loopback-only API between the native macOS shell and the embedded engine.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "43127"
security:
  - bearerAuth: []
paths:
...
  /v1/discovery/repository:
    post:
      operationId: discoverRepository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "200":
          description: Read-only discovery result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryDiscovery"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/projects:
    get:
      operationId: listProjects
      responses:
        "200":
          description: Projects known by this installation.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: importProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [repositoryPath]
              properties:
                repositoryPath:
                  type: string
                portableConfig:
                  $ref: "#/components/schemas/PortableProjectConfiguration"
      responses:
        "201":
          description: Imported draft project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getProject
      responses:
        "200":
          description: Project detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/projects/{projectId}/repositories/{repositoryId}/binding:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: repositoryId
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateRepositoryBinding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [path, bookmarkRef]
              properties:
                path:
                  type: string
                  minLength: 1
                bookmarkRef:
                  type: string
                  minLength: 1
                  maxLength: 300
      responses:
        "200":
          description: Project detail with the resolved Local Binding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/projects/{projectId}/configuration:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    put:
      operationId: replaceProjectConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [portableConfig, writeToRepository]
              properties:
                portableConfig:
                  $ref: "#/components/schemas/PortableProjectConfiguration"
                writeToRepository:
                  type: boolean
      responses:
        "200":
          description: Updated draft Project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/projects/{projectId}/bindings:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getProjectBindings
      responses:
        "200":
          description: Local Project Bindings without secret values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectBindings"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      operationId: replaceProjectBindings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectBindings"
      responses:
        "200":
          description: Updated Local Bindings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectBindings"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/projects/{projectId}/binding-candidates:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listProjectBindingCandidates
      responses:
        "200":
          description: Resources explicitly eligible for this Project's slots.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectResourceCandidate"
        default:
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid session bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: >-
        The request did not address the loopback interface. The engine binds
        127.0.0.1 only and rejects any other Host, per SYSTEM.md.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    RepositoryDiscovery:
      type: object
      additionalProperties: false
      required: [isGitRepository, suggested]
      properties:
        isGitRepository:
          type: boolean
        remoteUrl:
          type: [string, "null"]
        provider:
          type: [string, "null"]
        defaultBranch:
          type: [string, "null"]
        packageManager:
          type: [string, "null"]
        scripts:
          type: object
          additionalProperties:
            type: string
        suggested:
          type: object
    ProjectSummary:
      type: object
      additionalProperties: false
      required: [id, name, status, moduleCount]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [draft, valid, active, paused, invalid, degraded, archived]
        moduleCount:
          type: integer
          minimum: 0
        activeExecutions:
          type: integer
          minimum: 0
    ProjectDetail:
      # Not an allOf on ProjectSummary: that member's `additionalProperties: false`
      # would reject the detail's own `portableConfig`/`bindingStatus` properties,
      # making the schema unsatisfiable. It restates the summary properties instead.
      type: object
      additionalProperties: false
      required: [id, name, status, moduleCount, portableConfig, bindingStatus]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [draft, valid, active, paused, invalid, degraded, archived]
        moduleCount:
          type: integer
          minimum: 0
        activeExecutions:
          type: integer
          minimum: 0
        portableConfig:
          anyOf:
            - $ref: "#/components/schemas/PortableProjectConfiguration"
            - type: object
              description: Partial read-only discovery draft before configuration replacement.
        bindingStatus:
          type: object
          additionalProperties:
            type: object
            additionalProperties: false
            required: [path, accessible, bookmarkRef]
            properties:
              path:
                type: string
              accessible:
                type: boolean
              bookmarkRef:
                type: [string, "null"]
                description: Opaque reference to bookmark bytes owned by the macOS Shell.
...
    ProjectMetadata:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: string
          pattern: "^[a-z0-9][a-z0-9._-]*$"
          maxLength: 100
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 500
    ProjectRepositoryConfiguration:
      type: object
      additionalProperties: false
      required: [id, root, defaultBranch, remote]
      properties:
        id:
          type: string
          pattern: "^[a-z0-9][a-z0-9._-]*$"
        root:
          type: string
          const: "."
        defaultBranch:
          type: string
          minLength: 1
          maxLength: 200
        remote:
          type: string
          minLength: 1
          maxLength: 100
    ProjectSlotRequirement:
      type: object
      additionalProperties: false
      required: [requires]
      properties:
        requires:
          type: string
          pattern: "^[a-z0-9][a-z0-9._-]*$"
        optional:
          type: boolean
        description:
          type: string
    ProjectCommands:
      type: object
      additionalProperties: false
      properties:
        install: { type: string, minLength: 1 }
        lint: { type: string, minLength: 1 }
        typecheck: { type: string, minLength: 1 }
        test: { type: string, minLength: 1 }
        build: { type: string, minLength: 1 }
    ProjectGitConfiguration:
      type: object
      additionalProperties: false
      required: [branchPattern, commitStrategy, pushRemote]
      properties:
        branchPattern:
          type: string
          minLength: 1
          maxLength: 200
        commitStrategy:
          enum: [conventional, ticket-prefix, freeform]
        pushRemote:
          type: string
          minLength: 1
        allowForcePush:
          const: false
    ProjectWorkspaceConfiguration:
      type: object
      additionalProperties: false
      required: [strategy, maxConcurrentExecutions, retainOnFailureDays]
      properties:
        strategy:
          const: git-worktree
        maxConcurrentExecutions:
          type: integer
          minimum: 1
          maximum: 16
        retainOnFailureDays:
          type: integer
          minimum: 0
          maximum: 90
    ModuleInstanceConfiguration:
      type: object
      additionalProperties: false
      required: [instanceId, moduleId, enabled]
      properties:
        instanceId:
          type: string
          pattern: "^[a-z0-9][a-z0-9._-]*$"
        moduleId:
          type: string
          pattern: "^jarvis\\.module\\.[a-z0-9-]+$"
        enabled:
          type: boolean
        runtimeSlot:
          type: string
          pattern: "^[A-Za-z][A-Za-z0-9._-]*$"
        bindings:
          type: object
          additionalProperties:
            type: string
            pattern: "^[A-Za-z][A-Za-z0-9._-]*$"
        configuration:
          type: object
          additionalProperties: true
    PortableProjectConfiguration:
      type: object
      additionalProperties: false
      required: [apiVersion, kind, metadata, repositories, slots, commands, git, workspace, modules]
      properties:
        apiVersion:
          const: jarvis.dev/project/v1
        kind:
          const: Project
        metadata:
          $ref: "#/components/schemas/ProjectMetadata"
        repositories:
          type: array
          minItems: 1
          maxItems: 16
          items:
            $ref: "#/components/schemas/ProjectRepositoryConfiguration"
        slots:
          type: object
          minProperties: 1
          additionalProperties:
            $ref: "#/components/schemas/ProjectSlotRequirement"
        commands:
          $ref: "#/components/schemas/ProjectCommands"
        git:
          $ref: "#/components/schemas/ProjectGitConfiguration"
        workspace:
          $ref: "#/components/schemas/ProjectWorkspaceConfiguration"
        modules:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ModuleInstanceConfiguration"
    ProjectRepositoryBinding:
      type: object
      additionalProperties: false
      required: [path, bookmarkRef]
      properties:
        path:
          type: string
          minLength: 1
        bookmarkRef:
          type: [string, "null"]
          minLength: 1
    ProjectResourceCandidate:
      type: object
      additionalProperties: false
      required: [ref, kind, displayName, capabilities]
      properties:
        ref:
          type: string
        kind:
          type: string
          enum: [connection, runtime, mcp, module-instance, engine]
        displayName:
          type: string
        capabilities:
          type: array
          items:
            type: string
    ProjectSlotBinding:
      type: object
      additionalProperties: false
      required: [kind, ref]
      properties:
        kind:
          type: string
          enum: [connection, runtime, mcp, module-instance, engine]
        ref:
          type: string
          minLength: 1
          maxLength: 300
    ProjectBindings:
      type: object
      additionalProperties: false
      required: [apiVersion, kind, projectId, repositories, slots]
      properties:
        apiVersion:
          const: jarvis.dev/project-bindings/v1
        kind:
          const: ProjectBindings
        projectId:
          type: string
          pattern: "^[a-z0-9][a-z0-9._-]*$"
        repositories:
          type: object
          minProperties: 1
          additionalProperties:
            $ref: "#/components/schemas/ProjectRepositoryBinding"
        slots:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProjectSlotBinding"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            correlationId:
              type: string